Update to properly support Nitro users with 2000+ character messages; also a minor fix for display avatars in the conditional statement
New feature! NITRO support in server censoring... Now supporting reposting and editing messages in excess of 2000 characters to post full length messages from Nitro users. It uses the magic of embeds (which support content bodies of up to 4000 characters in the description field) to support long messages from Nitro users with the bot user.

Minor fix for the display avatar conditional statement. Technically, it should be checking for the display_avatar. Since avatar is the parent of display_avatar and passes through into it when no server-specific avatar is configured by the user, it would be working anyway.