NEW FEATURE: Requests for term removals; fix of blacklist modal from character overflow
Users can now submit requests for terms to be removed from the blacklists. Staff are notified in the log channel where an admin can easily approve or disapprove and provide a reason for such. Upon approval, the term is deleted from all blacklists that contain it. Additionally, in either case, the reporter is notified via DM. When a term is either already pending approval/disapproval or it has already been disapproved or approved in the past, the user is notified via ephemeral message of such and no additional request is sent.

Minor fix for character overflow that caused the modal window to not appear when editing blacklists. Had exceeded the 100-character limit for the modal placeholder in initially describing how to add URL blacklist terms. Reworded to fit all three examples.