Bug fix: Updated data structure to not mistakenly prune still-present censored messages in threads
This fixes a bug where the bot was pruning censored messages that occurred in threads. Since the bot could only see if a message occur in a parent channel and not a thread of the posting webhook, it assumed the censored_messaged had been deleted by its author when it had not. This fixes this by adding an optional thread_id entry that is used and will properly detect a message's existence in a thread if that message has not been deleted.

Note: If you have been self-hosting the bot, enter the database file "censored_messages.db" and add the thread_id column to your preexisting table like so:
sqlite3 server_configs/censored_messages.db
ALTER TABLE censored_messages ADD COLUMN thread_id INTEGER;

Verify the column was added successfully with this command:

.schema censored_messages

Then quit sqlite3 with this statement:

.quit